The rendering looks correct too - it'\''s drawing all the points in the path (line 3127-3130).
Wait, I think I may have misunderstood. Let me re-read what you said: "it makes it a shorter line as far as the dimensions".

Are you saying the **visual line on screen** is shorter, or the **measurement value/number** is smaller than it should be?

If the measurement value is wrong, there might be an issue with how `totalLength` is being calculated. Let me add some debug logging to verify. But first, can you confirm: when you draw a line that crosses itself, is the **number shown** (like "150mm") incorrect (too small)?